Рефакторинг большого виджета. Разбит на файлы
