+++ lib/UI/screens/components/custom_pickers/custom_date_picker.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+/// Пикер даты с кастомизированными цветами.
+class CustomDatePicker extends StatelessWidget {
+  final Widget child;
+  final Color? colorSchemePrimary;
+  final Color? colorSchemeOnPrimary;
+  final Color? colorSchemeOnSurface;
+  final Color? dialogBackgroundColor;
+  final ThemeData theme;
+
+  const CustomDatePicker({
+    Key? key,
+    required this.child,
+    required this.theme,
+    this.colorSchemePrimary,
+    this.colorSchemeOnPrimary,
+    this.colorSchemeOnSurface,
+    this.dialogBackgroundColor,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Theme(
+      data: theme.copyWith(
+        colorScheme: theme.colorScheme.copyWith(
+          primary: colorSchemePrimary,
+          onPrimary: colorSchemeOnPrimary,
+          onSurface: colorSchemeOnSurface,
+        ),
+        dialogBackgroundColor: dialogBackgroundColor,
+      ),
+      child: child,
+    );
+  }
+}
