+++ lib/UI/screens/components/empty_visiting_sight_list/base_empty_visiting_sight_list.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'package:places/helpers/app_strings.dart';
+
+/// Отображает информацию о пустом списке мест.
+///
+/// Имеет поля, которые необходимо переопределить в потомках:
+/// * [emptyInfo] - информация об отсутсвии записей;
+/// * [emptyIconPath] - иконка пустого списка.
+abstract class BaseEmptyVisitingList extends StatelessWidget {
+  static const double _iconSize = 64.0;
+
+  abstract final String emptyInfo;
+  abstract final String emptyIconPath;
+
+  const BaseEmptyVisitingList({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final textTheme = theme.textTheme;
+    final secondaryColor = theme.colorScheme.secondary.withOpacity(0.56);
+
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Image.asset(
+          emptyIconPath,
+          height: _iconSize,
+          width: _iconSize,
+        ),
+        const SizedBox(
+          height: 24,
+        ),
+        Text(
+          AppStrings.empty,
+          style: textTheme.subtitle1?.copyWith(
+            color: secondaryColor,
+          ),
+        ),
+        const SizedBox(
+          height: 8,
+        ),
+        Text(
+          emptyInfo,
+          style: textTheme.bodyText2?.copyWith(
+            color: secondaryColor,
+            height: 1.2,
+          ),
+          textAlign: TextAlign.center,
+        ),
+      ],
+    );
+  }
+}
