+++ lib/UI/screens/components/sight_card/sight_card_with_hover_ability.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+/// Карточка места с подсветкой в момент, когда над ней происходит перетаскивание другой карточки.
+/// Идентифицирует о возможности сделать дроп в эту область.
+class SightCardWithHoverAbility extends StatelessWidget {
+  final Widget sightCard;
+  final List<int?> candidateData;
+
+  const SightCardWithHoverAbility({
+    Key? key,
+    required this.sightCard,
+    required this.candidateData,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DecoratedBox(
+      decoration: BoxDecoration(
+        boxShadow: [
+          if (candidateData.isNotEmpty)
+            BoxShadow(
+              color: Theme.of(context).colorScheme.primary.withOpacity(0.4),
+              blurRadius: 2,
+              spreadRadius: 0.5,
+              offset: const Offset(0, -7),
+              blurStyle: BlurStyle.inner,
+            )
+          else
+            const BoxShadow(color: Colors.transparent),
+        ],
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: sightCard,
+    );
+  }
+}
